Attempt to simplify project layout
