--- cmd/frink/main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/uitml/frink/internal/commands"
-)
-
-func main() {
-	if err := commands.Execute(); err != nil {
-		os.Exit(1)
-	}
-}
